refactor: UserAttributesBuilder
